==> code/Analysis Script/CytokineMicrobiomeModel/03.CytokineMicrobiomeModel/01_DS_Code/02_model-validation.Rmd <==
---
title: "02_model-validation"
author: "Dan Spakowicz"
date: "9/25/2018"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(magrittr)
library(brms)
```

# Markdown syntax

```{r load data}
load("data/model-output.RData")

m <- brm
rm(brm)

# Change `days` back to numeric
m[["data"]]$days <- 
  m[["data"]]$days %>%
  as.character() %>%
  as.numeric()
```

# Model evaluation based on [the brms vignette https://cran.r-project.org/web/packages/brms/vignettes/brms_nonlinear.html]

```{r summary}
summary(m)
```

cite 
```{r plot}
# Plot population-level effects
plot(m, pars = "^b")
```

```{r marginal effects}
plot(marginal_effects(m), points = TRUE)
```

```{r posterior predictive}
pp_check(m)
```

```{r loo}
loo(m)
```

## Model evaluation based on [Mi et al., PLoS One, 2015. Goodness-of-Fit Tests and Model Diagnostics for Negative Binomial Regression of RNA Sequencing Data https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4365073/]
